Object conversions. Objects derive Hash,Eq,PartialEq. Add Entity.
